hehe

+++ cwt-spring/src/main/java/com/cwtsite/cwt/domain/ranking/view/model/RankingDto.kt
@@ -33,7 +33,7 @@ data class RankingDto(
             lastTournament = TournamentMinimalDto.toDto(ranking.lastTournament!!),
             lastPlace = ranking.lastPlace,
             lastDiff = ranking.lastDiff,
-            points = ranking.points.setScale(0, RoundingMode.HALF_UP),
+            points = ranking.points.setScale(0, RoundingMode.HALF_UP).toInt(),
             participations = ranking.participations,
             gold = ranking.gold,
             silver = ranking.silver,
